binary_sensor: fix Vacation Mode sensor state update
Since `absenceType` is immediately set to `VACATION` when configuring the Vacation Mode in the HomeaticIP app, even if the start time is in the future, the current system time is now compared with the`absenceStartTime` in addition to the `absenceType` in order to detect an active Vacation Mode. The start and end times are stored in ISO format in the sensor attributes for that purpose. `lastVacationTemperatur` is now used as  target temperature for the sensor because the Home object doesn't have a member `setPointTemperature`.