binary_sensor: PARTY mode not handled as absence vacation mode sensor
Party mode can be activated room by room in the HomematicIP app. For some reason (it is not entirely clear why), activating the party mode for a single room triggers a `HOME_CHANGED` event in which `absenceType`, `absenceStartTime`, and  `absenceEndTime` are set in the `INDOOR_CLIMATE` function group, even though it is only applied to one room by a `DEVICE_CHANGED` event. This should not be handled as a kind of an absence by the vacation mode sensor.