Add translations for service turn on with time and light effects (#235)
This pull request successfully adds translations for the switch_on_with_time service and for light effect states. The change to the maximum on_time in services.yaml correctly aligns the service definition with what appears to be an API limitation. The addition of _attr_translation_key in light.py is the correct approach for enabling these translations. I have one minor suggestion to improve the localization in the English translation file.

* Add translations for service turn on with time and light effects

* Add new Constraints: 0.1 - 16383 for service turn on with time as described in the API

* Add Code Review Changes

* Add Code Review Changes

* Add Code Review Changes