Fix/issue #164 climate visu (#242)
This pull request correctly implements the hvac_action property for the climate entity, which will allow for better visualization of the heating status in Home Assistant. The logic is sound and covers the necessary states. I have one minor comment regarding code style to improve maintainability.

* Add Method hvac_action

* Add HVACMode.OFF

* Code Improvements

So this is the fix for #164. I think the integration now does exactly what it did in the CCU world with Homematic IP Local. The one thing that isn’t implemented yet is cooling.